+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.1 - 2025-06-18
 * add possibility to iterate over pages when doing all requests instead of loading everything in memory
 * improve typing on mixins concerning map_model and raw_data parameters
+* update models (add create journal, create bank account, create ledger account)
 
 ## 0.2.0 - 2025-06-09
 * remove raw-data and x-chift-client-requestid from headers on uncessary endpoints